third_party/WebKit/LayoutTests/fast/dom/NodeList/nodelist-item-call-as-function.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description('This tests that items in a NodeList cannot be called indexed using [[Call]].');

var nodeList = document.getElementsByTagName('div');
shouldThrow("nodeList(0)");

</script>
</body>
</html>
